remove: drop SLP support and remove @psf/bch-js dependency
