+++ package.json
@@ -32,7 +32,6 @@
   "homepage": "https://github.com/paytaca/watchtower-cash-js#readme",
   "dependencies": {
     "@bitauth/libauth": "^3.0.0",
-    "@psf/bch-js": "^6.3.4",
     "axios": "^0.21.1",
     "bip39": "^3.1.0",
     "js-sha256": "^0.9.0"
