+++ src/address/index.ts
@@ -8,15 +8,8 @@ import {
   cashAddressTypeBitsToType,
   decodeBase58AddressFormat,
   cashAddressToLockingBytecode,
-  lockingBytecodeToBase58Address,
-  lockingBytecodeToCashAddress
+  lockingBytecodeToBase58Address
 } from "@bitauth/libauth";
-
-export enum SLPNetworkPrefix {
-  mainnet = "simpleledger",
-  testnet = "slptest",
-  regtest = "slpreg"
-}
 
 export default class Address {
   constructor (public address) {
@@ -67,7 +60,7 @@ export default class Address {
 
   toCashAddress () {
     const decoded = this._decodeCashaddr(this.address);
-    const network = Object.keys(CashAddressNetworkPrefix)[Object.values(CashAddressNetworkPrefix).findIndex(val => val === decoded.prefix)] ?? Object.keys(SLPNetworkPrefix)[Object.values(SLPNetworkPrefix).findIndex(val => val === decoded.prefix)];
+    const network = Object.keys(CashAddressNetworkPrefix)[Object.values(CashAddressNetworkPrefix).findIndex(val => val === decoded.prefix)];
     const type = decoded.type === CashAddressType.p2pkhWithTokens ? CashAddressType.p2pkh : (
       decoded.type === CashAddressType.p2shWithTokens ? CashAddressType.p2sh : decoded.type
     )
@@ -77,7 +70,7 @@ export default class Address {
   isCashAddress () {
     try {
       const decoded = this._decodeCashaddr(this.address);
-      return Object.values(CashAddressNetworkPrefix).includes(decoded.prefix as any) || Object.values(SLPNetworkPrefix).includes(decoded.prefix as any);
+      return Object.values(CashAddressNetworkPrefix).includes(decoded.prefix as any);
     } catch {
       return false;
     }
@@ -94,7 +87,7 @@ export default class Address {
 
   toTokenAddress () {
     const decoded = this._decodeCashaddr(this.address);
-    const network = Object.keys(CashAddressNetworkPrefix)[Object.values(CashAddressNetworkPrefix).findIndex(val => val === decoded.prefix)] ?? Object.keys(SLPNetworkPrefix)[Object.values(SLPNetworkPrefix).findIndex(val => val === decoded.prefix)];
+    const network = Object.keys(CashAddressNetworkPrefix)[Object.values(CashAddressNetworkPrefix).findIndex(val => val === decoded.prefix)];
     const type = decoded.type === CashAddressType.p2pkh ? CashAddressType.p2pkhWithTokens : (
       decoded.type === CashAddressType.p2sh ? CashAddressType.p2shWithTokens : decoded.type
     )
@@ -119,41 +112,6 @@ export default class Address {
     }
   }
 
-  isSLPAddress () {
-    try {
-      const decoded = this._decodeCashaddr(this.address);
-      return Object.values(SLPNetworkPrefix).includes(decoded.prefix as any);
-    } catch {
-      return false;
-    }
-  }
-
-  toSLPAddress () {
-    const decoded = this._decodeCashaddr(this.toCashAddress());
-    const network = Object.keys(CashAddressNetworkPrefix)[Object.values(CashAddressNetworkPrefix).findIndex(val => val === decoded.prefix)] ?? Object.keys(SLPNetworkPrefix)[Object.values(SLPNetworkPrefix).findIndex(val => val === decoded.prefix)];
-    return encodeCashAddress({ prefix: SLPNetworkPrefix[network], type: decoded.type, payload: decoded.payload }).address;
-  }
-
-  isMainnetSLPAddress () {
-    try {
-      const decoded = this._decodeCashaddr(this.address);
-      const network = Object.keys(SLPNetworkPrefix)[Object.values(SLPNetworkPrefix).findIndex(val => val === decoded.prefix)];
-      return Object.values(SLPNetworkPrefix).includes(decoded.prefix as any) && network === "mainnet";
-    } catch {
-      return false;
-    }
-  }
-
-  isTestnetSLPAddress () {
-    try {
-      const decoded = this._decodeCashaddr(this.address);
-      const network = Object.keys(SLPNetworkPrefix)[Object.values(SLPNetworkPrefix).findIndex(val => val === decoded.prefix)];
-      return Object.values(SLPNetworkPrefix).includes(decoded.prefix as any) && network === "testnet";
-    } catch {
-      return false;
-    }
-  }
-
   isValidBCHAddress (isChipnet = false) {
     if (isChipnet) {
       if (this.isLegacyAddress()) {
@@ -164,18 +122,6 @@ export default class Address {
     } else {
       return this.isCashAddress() || this.isTokenAddress();
     }
-    // const isBCHAddr = this.isCashAddress()
-    // if (isChipnet)
-    //   return isBCHAddr && this.isTestnetCashAddress()
-    // return isBCHAddr && this.isMainnetCashAddress()
-  }
-
-  isValidSLPAddress (isChipnet = false) {
-    return this.isSLPAddress()
-    // const isSLPAddr = this.isSLPAddress()
-    // if (isChipnet)
-    //   return isSLPAddr && this.isTestnetSLPAddress()
-    // return isSLPAddr && this.isMainnetSLPAddress()
   }
 
   private _decodeCashaddr(address: string): {
@@ -185,24 +131,20 @@ export default class Address {
     type: CashAddressType;
 } {
     let result: any;
-    // If legacy address convert first to cash address
     if (this.isLegacyAddress()) {
       const legacyResult = decodeBase58AddressFormat(address);
       if (typeof legacyResult === "string") {
         throw new Error(legacyResult);
       }
-      // Determine network from version byte
       const network = legacyResult.version === 0 ? CashAddressNetworkPrefix.mainnet : 
                       legacyResult.version === 5 ? CashAddressNetworkPrefix.mainnet :
                       CashAddressNetworkPrefix.testnet;
       const type = legacyResult.version === 0 || legacyResult.version === 111 ? CashAddressType.p2pkh : CashAddressType.p2sh;
       address = encodeCashAddress({ prefix: network, type, payload: legacyResult.payload }).address;
     }
-    // If the address has a prefix decode it as is
     if (address.includes(":")) {
       result = decodeCashAddressFormat(address);
     } else {
-      // otherwise, derive the network from the address without prefix
       result = decodeCashAddressFormatWithoutPrefix(address);
     }
